+++ src/components/List.tsx
@@ -1,8 +1,8 @@
 import { Text } from "ink"
 import React from "react"
 
-import { useGames } from "../../gql/useGames.js"
-import { useMe } from "../../gql/useMe.js"
+import { useGames } from "../gql/useGames.js"
+import { useMe } from "../gql/useMe.js"
 
 export function List() {
   const { me } = useMe()
